Increase changes of registration checks working
Right now this script regularly fails for me during the registration step, however it seems that this is just because registration is taking longer than expected. When I pause and run the check-registration command myself it works.

What I've done here is-

* Add a five second pause before the very first check.
* Increased the sleep period from three to five seconds.
* Increased the number of checks from three to five.

This changes the total waiting period from a maximum of six seconds to 25 seconds, which seems to be more than enough time.

One change I did *not* make, because it would break backwards compatibility, is to change the exist code to fail when the registration doesn't work. I think this should be done at some point- we're running this script as part of a build process and I think if the agent can't register to Vanta then the build should fail.