Remove Rosetta check from macOS install script
As of agent version 1.8.5, the agent is a universal binary, so Rosetta is no longer needed. Additionally, from experience earlier today, the Rosetta plist may no longer be where this script expects it to be, due to successive revisions to Rosetta itself, so the script might fail even if Rosetta is installed.